Add missing documentation, refactor naming
